+++ DocumentFormat.OpenXml/OfficeAvailabilityAttribute.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) Microsoft Open Technologies, Inc.  All rights reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using System;
+
+namespace DocumentFormat.OpenXml
+{
+    /// <summary>
+    /// Defines an OfficeAvailabilityAttribute class to indicate whether the property is available in a specific version of an Office application.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Property | AttributeTargets.Class | AttributeTargets.Field)]
+    public sealed class OfficeAvailabilityAttribute : Attribute
+    {
+        /// <summary>
+        /// Gets the Office version of the available property.
+        /// </summary>
+        public FileFormatVersions OfficeVersion { get; internal set; }
+
+        /// <summary>
+        /// Initializes a new instance of the OfficeAvailabilityAttribute class.
+        /// </summary>
+        /// <param name="officeVersion">The Office version where this class or property is available. 
+        /// If there is more than one version, use bitwise OR to specify multiple versions.</param>
+        public OfficeAvailabilityAttribute(FileFormatVersions officeVersion)
+        {
+            OfficeVersion = officeVersion;
+        }
+    }
+}
