+++ DocumentFormat.OpenXml/OfficeVersions.cs
@@ -4,33 +4,6 @@ using System;
 
 namespace DocumentFormat.OpenXml
 {
-    /// <summary>
-    /// Defines the Office Open XML file format version.
-    /// </summary>
-    [Flags]
-    public enum FileFormatVersions
-    {
-        /// <summary>
-        /// Represents an invalid value which will cause an exception.
-        /// </summary>
-        None = 0,
-
-        /// <summary>
-        /// Represents Microsoft Office 2007.
-        /// </summary>
-        Office2007 = 0x1,
-
-        /// <summary>
-        /// Represents Microsoft Office 2010.
-        /// </summary>
-        Office2010 = 0x2,
-
-        /// <summary>
-        /// Represents Microsoft Office 2013.
-        /// </summary>
-        Office2013 = 0x4
-    }
-
     internal static class OfficeVersions
     {
         /// <summary>
@@ -78,28 +51,6 @@ namespace DocumentFormat.OpenXml
 
                 throw new ArgumentOutOfRangeException(parameterName, message);
             }
-        }
-    }
-
-    /// <summary>
-    /// Defines an OfficeAvailabilityAttribute class to indicate whether the property is available in a specific version of an Office application.
-    /// </summary>
-    [AttributeUsage(AttributeTargets.Property | AttributeTargets.Class | AttributeTargets.Field)]
-    public sealed class OfficeAvailabilityAttribute : Attribute
-    {
-        /// <summary>
-        /// Gets the Office version of the available property.
-        /// </summary>
-        public FileFormatVersions OfficeVersion { get; internal set; }
-
-        /// <summary>
-        /// Initializes a new instance of the OfficeAvailabilityAttribute class.
-        /// </summary>
-        /// <param name="officeVersion">The Office version where this class or property is available. 
-        /// If there is more than one version, use bitwise OR to specify multiple versions.</param>
-        public OfficeAvailabilityAttribute(FileFormatVersions officeVersion)
-        {
-            OfficeVersion = officeVersion;
         }
     }
 }
